Feat: 검증 추가, 결과를 생성하여 반환하는 기능 구현
